[UserDidAssociation] service added.
