+++ src/services/api/userDidAssociation/userDidAssociation.service.ts
@@ -2,18 +2,18 @@ import { ServiceAddons } from '@feathersjs/feathers';
 
 import { Application } from '../../../declarations';
 import { CredentialEntity } from '../../../entities/Credential';
-import { hooks } from './userDid.hooks';
-import { CredentialRequestService } from './userDid.class';
+import { hooks } from './userDidAssociation.hooks';
+import { UserDidAssociationService } from './userDidAssociation.class';
 
 // add this service to the service type index
 declare module '../../../declarations' {
   interface ServiceTypes {
-    credentialRequest: CredentialRequestService & ServiceAddons<CredentialEntity>;
+    userDidAssociation: UserDidAssociationService & ServiceAddons<any>;
   }
 }
 
 export default function (app: Application): void {
-  app.use('/credentialRequest', new CredentialRequestService({}, app));
-  const service = app.service('credentialRequest');
+  app.use('/userDidAssociation', new UserDidAssociationService({}, app));
+  const service = app.service('userDidAssociation');
   service.hooks(hooks);
 }
